Remove 15 more no-op return-type casts in four object DLLs
Same shape as the main wave; these four TUs were held back by the
freshness guard. All four objects byte-identical; main.dol: OK.

+++ src/dlls/objects/470/470.c
@@ -81,8 +81,8 @@ void dll_1D6_update(GameObject* obj) {
     if ((state->flags & DLL1D6_STATE_FLAG_DOWN_PHASE) != 0) {
         if ((state->flags & DLL1D6_STATE_FLAG_BOB_ACTIVE) == 0) {
             state->flags |= DLL1D6_STATE_FLAG_BOB_ACTIVE;
-            state->bobPhase = (f32)(int)randomGetRange(20, 40);
-            state->bobRate = (f32)(int)randomGetRange(6, 10) / 20.0f;
+            state->bobPhase = (f32)randomGetRange(20, 40);
+            state->bobRate = (f32)randomGetRange(6, 10) / 20.0f;
         }
         state->downTimer -= framesThisStep;
         state->dizzyTimer = state->dizzyTimer - framesThisStep;
@@ -178,10 +178,10 @@ void dll_1D6_update(GameObject* obj) {
     if ((state->flags & DLL1D6_STATE_FLAG_BOB_ACTIVE) != 0) {
         state->bobPhase = state->bobRate * timeDelta + state->bobPhase;
         if (state->bobPhase > 40.0f) {
-            state->bobRate = -(f32)(int)randomGetRange(6, 10) / 20.0f;
+            state->bobRate = -(f32)randomGetRange(6, 10) / 20.0f;
             state->bobPhase = 40.0f;
         } else if (state->bobPhase < 20.0f) {
-            state->bobRate = (f32)(int)randomGetRange(6, 10) / 20.0f;
+            state->bobRate = (f32)randomGetRange(6, 10) / 20.0f;
             state->bobPhase = 20.0f;
         }
     }
